FEATURE: Adding SideMenuHeader
