+++ src/components/ui/molecules/Dropdown/Dropdown.styles.js
@@ -21,13 +21,14 @@ export const Wrapper = styled.ul`
   flex-direction: column;
   height: ${({ height }) => height && height};
   justify-content: flex-start;
-  left: ${({ left }) => (left ? "0.8rem" : null)};
   overflow: hidden;
   padding: 0.8rem;
   position: absolute;
-  right: ${({ left }) => (left ? null : "0.8rem")};
+  /* right: ${({ left }) => (left ? null : "0.8rem")}; */
+  right: 0.8rem;
   top: calc(100px + 0.8rem);
   transition: all 100ms linear, height 500ms ease;
   width: 250px;
   z-index: 1;
+  left: ${({ left }) => (left ? "0.8rem" : null)};
 `;
