+++ src/components/ui/molecules/index.js
@@ -1,3 +1,4 @@
 export { default as Dropdown } from "./Dropdown/Dropdown.component";
 export { default as Navigation } from "./Navigation/Navigation.component";
+export { default as SideMenuHeader } from "./SideMenuHeader/SideMenuHeader.component";
 export { default as UserIndicator } from "./UserIndicator/UserIndicator.component";
